Put links to recipes rather than wikipedia
Links to wikipedia articles is kind of douchy, instead this change is to put links to the actual recipes for my own sake and then also to give credit to the people with the actual recipes. I still need to remember which recipes I liked for Jollof rice, spaghetti with meatballs and gumbo. Also reminds me that I should start a section with my favourite hot sauces from different places!